requirements.txt
numpy==2.1.3    
opencv-python==4.10.0.84
pillow==11.0.0   
pip==24.2     
tk==0.1.0